- added message window for script output (DefoScriptWidget) - GUI controls are now disabled when script is running - a new measurement directory can be created from within   a script

+++ defo/DefoScriptableJulabo.cc
@@ -13,10 +13,10 @@ DefoScriptableJulabo::DefoScriptableJulabo(
 {
   connect(this, SIGNAL(changeP(double)),
           julaboModel_,SLOT(setProportionalValue(double)));
-  connect(this, SIGNAL(changeTv(double)),
-          julaboModel_,SLOT(setIntegralValue(double)));
-  connect(this, SIGNAL(changeTd(double)),
-          julaboModel_,SLOT(setDifferentialValue(double)));
+  connect(this, SIGNAL(changeTv(int)),
+          julaboModel_,SLOT(setIntegralValue(int)));
+  connect(this, SIGNAL(changeTd(int)),
+          julaboModel_,SLOT(setDifferentialValue(int)));
   
   connect(this, SIGNAL(switchCirculator(bool)),
           julaboModel_, SLOT(setCirculatorEnabled(bool)));
@@ -25,7 +25,7 @@ DefoScriptableJulabo::DefoScriptableJulabo(
           julaboModel_, SLOT(setWorkingTemperatureValue(double)));
 }
 
-void DefoScriptableJulabo::setPID( double p, double tv, double td ) {
+void DefoScriptableJulabo::setPID( double p, int tv, int td ) {
 
   QMutexLocker locker(&mutex_);
